+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 更新日志
 
+#### 2.0.8 (2021-09-23)
+
+* 增加显示关闭组件监听函数```onShow```，主要用于js方式调用，```Function(isShow: boolean)```
+* 更新README
+
 #### 2.0.7 (2021-07-27)
 
 * Reset transform when modal is visible
@@ -35,6 +40,17 @@
 * 升级为vue@3.x版本
 * 基于vite@2.x开发
 * vue@2.x版本请使用 [![vue-img-viewr@1.0.4](https://img.shields.io/badge/npm%20vue--img--viewr-v1.0.4-blue)](https://www.npmjs.com/package/vue-img-viewr/v/1.0.4)， Github v1.x地址 [![github 1.x](https://img.shields.io/badge/github%20vue--img--viewr-1.x-green)](https://github.com/jekorx/vue-img-viewr/tree/1.x)
+
+#### 1.0.6 (2021-09-23)
+
+* 增加显示关闭组件监听函数```onShow```，主要用于js方式调用，```Function(isShow: boolean)```
+* 更新README
+
+#### 1.0.5 (2021-07-27)
+
+* Reset transform when modal is visible
+* body lock use classList
+* README
 
 #### 1.0.4 (2021-01-21)